Refactor cli.go
As cli.go is currently ~1800 lines long, a refactor is in order.
This commit reduces the size of cli.go by splitting each command
handler off to its own file. These handler files will also contain any
helper functions used exclusively by that handler. Shared helper
functions are instead kept in cli.go. Finally, type and variable
definitions have been moved to value_definitions.go.
21 new files have been created, and now
the longest file is only ~400 lines long.
